Refactor `WaitingTimeTracker` to remove `Analytics` dependency so that it can be moved to WooFoundation. `AppStartupWaitingTimeTracker` is now a wrapper of `WaitingTimeTracker`.
